Book Store with Documentation
Design and implement a Java program to create a system that will keep track of books in a local bookstore. The program will maintain a current list of books that can be purchased in a bookstore and ask the user if they would like to order the book if it is not in stock. Book data for the program includes the author’s first and last and an array of books written by the author which will contain name of book, isbn number, price and genre. Each author can have a maximum of five books.